move test dataobject to tests folder / first test of task
